requirements.txt
streamlit
torch
torchvision
pillow
pandas
scikit-learn
matplotlib